+++ src/dir_empm/interp_x_c_to_k_p_xxnufft.py
@@ -128,15 +128,15 @@ def interp_x_c_to_k_p_xxnufft(
     #end;%if flag_u_vs_c==0;
 
     n_w_max = int(torch.sum(n_w_).item()) ;
-    k0_wk_ = torch.zeros(n_w_max).to(dtype=torch.float64) ;
-    k1_wk_ = torch.zeros(n_w_max).to(dtype=torch.float64) ;
+    k0_wk_ = torch.zeros(n_w_max).to(dtype=torch.float32) ;
+    k1_wk_ = torch.zeros(n_w_max).to(dtype=torch.float32) ;
 
     S_k_p_wk_ = torch.zeros(n_w_max, dtype=torch.complex128) ;
     n_w_max = int(torch.max(n_w_).item()); n_w_sum = int(torch.sum(n_w_).item());
     if numel_unique(n_w_)==1:
-        k0_wk__ = torch.zeros(mtr((n_w_sum,n_k_p_r))).to(dtype=torch.float64);
-        k1_wk__ = torch.zeros(mtr((n_w_sum,n_k_p_r))).to(dtype=torch.float64);
-        omega_w_ = 2*pi*torch.arange(n_w_max).to(dtype=torch.float64)/np.maximum(1,n_w_max);
+        k0_wk__ = torch.zeros(mtr((n_w_sum,n_k_p_r))).to(dtype=torch.float32);
+        k1_wk__ = torch.zeros(mtr((n_w_sum,n_k_p_r))).to(dtype=torch.float32);
+        omega_w_ = 2*pi*torch.arange(n_w_max).to(dtype=torch.float32)/np.maximum(1,n_w_max);
         k0_wk__ = mmmm(torch.reshape(torch.cos(omega_w_),mtr((n_w_max,1))),torch.reshape(2*pi*k_p_r_,mtr((1,n_k_p_r))));
         k0_wk_ = k0_wk__.ravel();
         k1_wk__ = mmmm(torch.reshape(torch.sin(omega_w_),mtr((n_w_max,1))),torch.reshape(2*pi*k_p_r_,mtr((1,n_k_p_r))));
